- refactor: extract tor bootstrap logic into independent component and use from the node facade
