+++ androidNode/src/androidMain/kotlin/network/bisq/mobile/android/node/service/bootstrap/NodeApplicationBootstrapFacade.kt
@@ -16,6 +16,7 @@ import kotlinx.coroutines.CancellationException
 import network.bisq.mobile.android.node.AndroidApplicationService
 import network.bisq.mobile.android.node.service.tor.TorIntegrationService
 import network.bisq.mobile.android.node.service.tor.TorService
+import network.bisq.mobile.android.node.service.tor.TorBootstrapOrchestrator
 import network.bisq.mobile.domain.service.bootstrap.ApplicationBootstrapFacade
 import network.bisq.mobile.domain.service.network.ConnectivityService
 import network.bisq.mobile.i18n.i18n
@@ -30,13 +31,12 @@ interface ApplicationServiceInitializationCallback {
 class NodeApplicationBootstrapFacade(
     private val applicationService: AndroidApplicationService.Provider,
     private val connectivityService: ConnectivityService,
-    private val torIntegrationService: TorIntegrationService
-) : ApplicationBootstrapFacade() {
+    private val torIntegrationService: TorIntegrationService,
+    private val torBootstrapOrchestrator: TorBootstrapOrchestrator
+) : ApplicationBootstrapFacade(), TorBootstrapOrchestrator.TorBootstrapCallback {
 
     private val applicationServiceState: Observable<State> by lazy { applicationService.state.get() }
-    private var torMonitoringJob: Job? = null
     private var applicationServiceStatePin: Pin? = null
-    private var torBootstrapComplete = CompletableDeferred<Boolean>()
     private var bootstrapSuccessful = false
     private var initializationCallback: ApplicationServiceInitializationCallback? = null
     private lateinit var serverSocket: ServerSocket
@@ -48,6 +48,44 @@ class NodeApplicationBootstrapFacade(
 
     fun setInitializationCallback(callback: ApplicationServiceInitializationCallback) {
         this.initializationCallback = callback
+    }
+
+    // TorBootstrapOrchestrator.TorBootstrapCallback implementation
+    override fun onTorStateChanged(message: String, progress: Float) {
+        setState(message)
+        setProgress(progress)
+    }
+
+    override fun onTorReady(socksPort: Int) {
+        log.i { "🚀 Bootstrap: Tor is ready - proceeding with application bootstrap" }
+
+        // CRITICAL: Configure Bisq for external Tor IMMEDIATELY when Tor becomes ready
+        // This must happen BEFORE any Bisq services are initialized
+        configureBisqForExternalTorEarly(socksPort)
+
+        // Proceed with application bootstrap
+        proceedWithApplicationBootstrap()
+    }
+
+    override fun onTorTimeout() {
+        log.w { "⚠️ Bootstrap: Tor timeout - proceeding with application bootstrap anyway" }
+        setState("Tor timeout - Starting Bisq...")
+        setProgress(0.25f)
+        launchIO {
+            delay(2000) // Show timeout message
+            proceedWithApplicationBootstrap()
+        }
+    }
+
+    override fun onTorError(exception: Exception) {
+        log.e(exception) { "❌ Bootstrap: Tor initialization failed" }
+        log.w { "⚠️ Bootstrap: Proceeding without Tor - users can enable it in settings" }
+        setState("Tor failed - Starting Bisq...")
+        setProgress(0.25f)
+        launchIO {
+            delay(2000)
+            proceedWithApplicationBootstrap()
+        }
     }
 
     override fun activate() {
@@ -87,209 +125,11 @@ class NodeApplicationBootstrapFacade(
      * This ensures Tor is fully ready before any network-dependent services start
      */
     private fun initializeAndWaitForTor() {
-        log.i { "🚀 Bootstrap: Initializing embedded Tor daemon and waiting for ready state..." }
-
-        // TODO i18n
-        setState("Initializing Tor daemon...")
-        setProgress(0.05f) // Very small progress to show we've started
-
-        setupTorStateObserver()
-
-        torMonitoringJob = torBootstrapScope.launch {
-            try {
-                // Start Tor with retry logic
-                torIntegrationService.initializeAndStart(
-                    maxRetries = 3,
-                    retryDelayMs = 5000
-                )
-
-                log.i { "✅ Bootstrap: Tor initialization started - waiting for ready state..." }
-                setState("Starting Tor daemon...")
-                setProgress(0.1f)
-
-                // Add periodic status checks during the wait
-                val statusCheckJob = launch {
-                    repeat(12) { // Check every 5 seconds for 60 seconds total
-                        delay(5000)
-                        setProgress(progress.value + 0.1f)
-                        val currentState = torIntegrationService.torState.value
-                        val currentPort = torIntegrationService.socksPort.value
-                        log.i { "🔍 Bootstrap: Tor status check - State: $currentState, Port: $currentPort" }
-
-                        // Update UI based on current state
-                        when (currentState) {
-                            TorService.TorState.STARTING -> setState("Starting Tor daemon...")
-                            TorService.TorState.BOOTSTRAPPING -> setState("Tor connecting to network...")
-                            TorService.TorState.READY -> {
-                                if (currentPort != null) {
-                                    setState("Tor ready - Starting Bisq...")
-                                } else {
-                                    setState("Tor almost ready...")
-                                }
-                            }
-                            else -> { /* Keep current state */ }
-                        }
-                    }
-                }
-
-                // Wait for Tor to become ready with timeout
-                val torReady = withTimeoutOrNull(60000) { // 60 second timeout
-                    torBootstrapComplete.await()
-                }
-
-                statusCheckJob.cancel() // Stop status checks
-
-                if (torReady == true) {
-                    log.i { "🚀 Bootstrap: Tor is ready - proceeding with application bootstrap" }
-                    setState("Tor ready - Starting Bisq...")
-                    setProgress(0.25f)
-                    delay(1000) // Show message briefly
-                    proceedWithApplicationBootstrap()
-                } else {
-                    log.w { "⚠️ Bootstrap: Tor timeout after 60 seconds" }
-
-                    // Debug and try to fix Tor status
-                    torIntegrationService.debugAndFixTorStatus()
-
-                    // Wait a bit for the debug fix to take effect
-                    delay(2000)
-
-                    // Final status check before giving up
-                    val finalState = torIntegrationService.torState.value
-                    val finalPort = torIntegrationService.socksPort.value
-                    log.w { "⚠️ Bootstrap: Final Tor status after debug - State: $finalState, Port: $finalPort" }
-
-                    // If Tor is actually ready but we missed the signal, proceed anyway
-                    if (finalState == TorService.TorState.READY && finalPort != null) {
-                        log.i { "🚀 Bootstrap: Tor was actually ready after debug - proceeding with bootstrap" }
-                        setState("Tor ready - Starting Bisq...")
-                        setProgress(0.25f)
-                        delay(1000)
-                        proceedWithApplicationBootstrap()
-                    } else {
-                        setState("Tor timeout - Starting Bisq...")
-                        setProgress(0.25f)
-                        delay(2000) // Show timeout message
-                        proceedWithApplicationBootstrap()
-                    }
-                }
-
-            } catch (e: Exception) {
-                if (e is CancellationException) {
-                    log.d { "🔄 Bootstrap: Tor initialization cancelled (normal during deactivation)" }
-                    return@launch
-                }
-                log.e(e) { "❌ Bootstrap: Failed to start Tor initialization" }
-                log.w { "⚠️ Bootstrap: Proceeding without Tor - users can enable it in settings" }
-                setState("Tor failed - Starting Bisq...")
-                setProgress(0.25f)
-                delay(2000)
-                proceedWithApplicationBootstrap()
-            }
-        }
+        log.i { "🚀 Bootstrap: Delegating Tor initialization to TorBootstrapOrchestrator..." }
+        torBootstrapOrchestrator.initializeAndWaitForTor(this, jobsManager)
     }
 
-    /**
-     * Set up observer for Tor state changes to detect when it becomes ready
-     */
-    private fun setupTorStateObserver() {
-        // Launch a coroutine to collect from the StateFlow
-        // Use a separate job that completes once Tor is ready
-        jobsManager.addJob(torBootstrapScope.launch {
-            try {
-                var shouldContinue = true
 
-                // Monitor both Tor state and SOCKS port simultaneously
-                torIntegrationService.torState.collect { torState ->
-                    if (!shouldContinue) return@collect
-
-                    log.i { "🔍 Bootstrap: Tor state changed to: $torState" }
-
-                    when (torState) {
-                        TorService.TorState.STARTING -> {
-                            setState("Starting Tor daemon...")
-                            setProgress(0.1f)
-                        }
-                        TorService.TorState.BOOTSTRAPPING -> {
-                            setState("Tor connecting to network...")
-                            setProgress(0.15f)
-                        }
-                        TorService.TorState.READY -> {
-                            // Check if we have both READY state and SOCKS port
-                            checkTorReadiness()
-                            shouldContinue = false // Stop collecting once we reach READY
-                        }
-                        TorService.TorState.ERROR -> {
-                            log.e { "❌ Bootstrap: Tor encountered an error" }
-                            // Complete with failure
-                            if (!torBootstrapComplete.isCompleted) {
-                                torBootstrapComplete.complete(false)
-                            }
-                            // Stop collecting on error
-                            shouldContinue = false
-                        }
-                        else -> {
-                            // Other states, continue waiting
-                        }
-                    }
-                }
-            } catch (e: Exception) {
-                if (e is CancellationException) {
-                    log.d { "🔄 Bootstrap: Tor state monitoring cancelled (normal during deactivation)" }
-                } else {
-                    log.e(e) { "❌ Bootstrap: Error in Tor state observer" }
-                    if (!torBootstrapComplete.isCompleted) {
-                        torBootstrapComplete.complete(false)
-                    }
-                }
-            }
-        })
-
-        // Also monitor SOCKS port separately to handle cases where port is available after READY state
-        jobsManager.addJob(torBootstrapScope.launch {
-            try {
-                torIntegrationService.socksPort.collect { socksPort ->
-                    log.i { "🔍 Bootstrap: SOCKS port changed to: $socksPort" }
-
-                    // If we have both READY state and SOCKS port, complete bootstrap
-                    if (socksPort != null && torIntegrationService.torState.value == TorService.TorState.READY) {
-                        checkTorReadiness()
-                    }
-                }
-            } catch (e: Exception) {
-                if (e is CancellationException) {
-                    log.d { "🔄 Bootstrap: SOCKS port monitoring cancelled (normal during deactivation)" }
-                } else {
-                    log.e(e) { "❌ Bootstrap: Error in SOCKS port observer" }
-                }
-            }
-        })
-    }
-
-    /**
-     * Check if Tor is fully ready (both state READY and SOCKS port available)
-     */
-    private fun checkTorReadiness() {
-        val torState = torIntegrationService.torState.value
-        val socksPort = torIntegrationService.socksPort.value
-
-        log.i { "🔍 Bootstrap: Checking Tor readiness - State: $torState, SOCKS Port: $socksPort" }
-
-        if (torState == TorService.TorState.READY && socksPort != null) {
-            log.i { "🚀 Bootstrap: Tor fully ready with SOCKS port: $socksPort" }
-
-            // CRITICAL: Configure Bisq for external Tor IMMEDIATELY when Tor becomes ready
-            // This must happen BEFORE any Bisq services are initialized
-            configureBisqForExternalTorEarly(socksPort)
-
-            // Complete the deferred to signal Tor is ready
-            if (!torBootstrapComplete.isCompleted) {
-                torBootstrapComplete.complete(true)
-            }
-        } else {
-            log.i { "⏳ Bootstrap: Tor not fully ready yet - waiting for both READY state and SOCKS port" }
-        }
-    }
 
 
 
@@ -1026,25 +866,8 @@ class NodeApplicationBootstrapFacade(
 
 
     override fun deactivate() {
-        if (torBootstrapComplete.isCompleted) {
-            // Only cancel Tor monitoring if bootstrap was not successful
-            // If bootstrap was successful, let Tor continue running
-            if (!bootstrapSuccessful) {
-                log.w { "⚠️ Bootstrap failed - cancelling Tor monitoring" }
-                torMonitoringJob?.cancel()
-                torMonitoringJob = null
-
-                // Reset the CompletableDeferred only if bootstrap failed
-                if (!torBootstrapComplete.isCompleted) {
-                    torBootstrapComplete.cancel()
-                }
-                torBootstrapComplete = CompletableDeferred()
-            } else {
-                log.i { "✅ Bootstrap successful - keeping Tor monitoring active" }
-            }
-        } else {
-            log.w { "⚠️ Bootstrap not completed - skipping cancellation of Tor monitoring" }
-        }
+        // Delegate Tor monitoring cancellation to the orchestrator
+        torBootstrapOrchestrator.cancelTorMonitoring(bootstrapSuccessful)
 
         applicationServiceStatePin?.unbind()
         applicationServiceStatePin = null
